+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Revision history for static-ls
 
-## 0.1.0.0 -- YYYY-mm-dd
+## 0.1.0.0 -- 2023-04-13
 
-* First version. Released on an unsuspecting world.
+* Initial release supporting
+  * find references globally when hie files and hiedb are available
+  * type on hover when hie files are available
+  * find definition locally and globally when hie files and hiedb are available